+++ tests/testthat/test-joins.r
@@ -303,3 +303,28 @@ test_that("JoinFactorFactorVisitor_SameLevels preserve levels order (#675)",{
   expect_equal( levels(res$g2), levels(output$g2)) 
 })
 
+test_that("joins coerce factors with different levels to character (#684)", {
+  d1 <- data_frame( a = factor( c("a", "b", "c" ) ) )
+  d2 <- data_frame( a = factor( c("a", "e" ) ) )
+  expect_warning( { res <- inner_join( d1, d2 ) })
+  expect_is( res$a, "character" )
+  
+  # different orders
+  d2 <- d1
+  attr( d2$a, "levels" ) <- c("c", "b", "a" )
+  expect_warning( { res <- inner_join( d1, d2 ) })
+  expect_is( res$a, "character" )
+  
+})
+
+test_that("joins between factor and character coerces to character with a warning (#684)", {
+  d1 <- data_frame( a = factor( c("a", "b", "c" ) ) )
+  d2 <- data_frame( a = c("a", "e" ) )
+  expect_warning( { res <- inner_join( d1, d2 ) })
+  expect_is( res$a, "character" )
+  
+  expect_warning( { res <- inner_join( d2, d1 ) })
+  expect_is( res$a, "character" )
+    
+})
+