+++ NEWS.md
@@ -39,6 +39,10 @@
   
 * `DataFrameVisitors` handles data frames as columns of data frames, recursively. (#602)
 
+* joining factors whose levels are not identical (same levels, same order) generates
+  a warning and coerces to character (#684). Same when joining a character vector
+  with a factor and vice versa. 
+
 # dplyr 0.3.0.1
 
 * Fixed problem with test script on Windows.